(mod/api)adding node_controller
currently, node CRUD does not work as we still need to add the address
CRUD as well, in order to respect the Foreign Key constraint.

this will be my next task, then I will have to update the tests to
reflect those changes.

then, we can move on to graphql testing, in order to finally proceed to
the benchmarks.