Release 0.3.0 (#57)
* (add/naboo)GraphQL-vs-REST benchmarks, Domain controllers, & utils module (#56)

* (mod/ci)bumping erlang version to otp25.0.4

would not compile on my machine using 24.1.7

also updated docker-compose syntax specs to 3.9 since this is not
breaking for us

* (mod/api)adding node_controller

currently, node CRUD does not work as we still need to add the address
CRUD as well, in order to respect the Foreign Key constraint.

this will be my next task, then I will have to update the tests to
reflect those changes.

then, we can move on to graphql testing, in order to finally proceed to
the benchmarks.

* (fix/naboo)assocs between nodes and addresses

foreign-key setup was not properly working between nodes and addresses.
for now, we are using a one-to-one relationship but this might change in
the future; some tests have been fixed as well, but some are still
crashing. dont know why yet.

* (mod/build)adding shell rule

shell rule shall provide an interactive IEx shell with the project modules
loaded

* (fix/api)unit tests for domains

- fixed unit tests for node & address controllers
- fixed some issue with migrations
- enhanced error messages for session_controller & address_controller

* (fix/api)improving session tests
* (mod/build)bump postgrex to 0.16
* (mod/build)dependency update utility

to better check if dependencies are up to date, I added a Makefile rule
to check for them against hex.pm

updated swagger provider as a result, pretty nice :)~

* (mod/api)tests for Domain GraphQL queries/mutations

this is the final step before running benchmarks! now i just have to
edit the locustfile to run the queries using both methods, and save the
results, and then we will finally be done.

* (fix/api)filtering input keys when creating node

depending on the input, we can have strings represented as string or an
atom, so we filter that at the module level.

* (add/docs)graphql vs REST benchmarks

this is the first of probably many benchmarks to understand and
determine which may be best in terms of performance, depending on the
scenario

* (mod/git)adding __pycache__ to .gitignore

* (add/naboo)utils module

a module which will comports random _helper_libraries

For now, it holds a Jason tuple encoder, and a method to convert
bitstring key-maps to atom key-maps

* (fix/naboo)unit tests & input handling on REST queries


Co-authored-by: winds0r <w.nds0r@protonmail.com>