(mod/api)account to domain foreign keys (#65)
* (mod/naboo)adding foreign keys between accounts & nodes

this is the first part of the work we have to do to have nodes &
addresses actually being stored by users. next up:

- encode user id in jwt -> when they create a node, cast their user_id
  with it
(figuring out a way to do this for graphQL as well would be nice, but it
is not mandatory at this point)
- edit changesets to cast the association
- add functions to query all of the necessary data depending on a user
  id, but not touch the current one because it might slow down
  performances overall
- edit & fix tests to reflect these changes

* (mod/ci)specifying DATABASE_SSL & STATIC_URL

AFAIK, these values are not _actually_ used for now, but the work done
to fix release mode has imposed us to use fetch_env! for everything,
which breaks startup of the app is the env value is not set

* (fix/test)specifying DATABASE_POOL_SIZE

* (fix/api)improving account creation error checking

-> changed the way changeset return errors when changeset is invalid, so
that API is more consistent
-> enforcing unique constraints on account creation (not as a business
rule)
-> removing password_confirmation field from account creation, as it
should be done in the front-end
-> checking wether email matches basic email regex

* (mod/build)fixing env values in prod docker-compose

* (mod/api)refactoring views for preloaded data

To not repeat code, we can just call a view from another one, when we
know there is a preloaded query. this is much cleaner, i like this
commit very much.

* (mod/api)setting up foreign-key relationships between nodes & accounts

we can now query an account with its related nodes, this is still
untested, but we need to change some stuff in the node_controller before
testing this part of the feature

* (mod/api)fixed assocs between users & domains