Release 0.4.0 (#66)
* (fix/build)release mode would not properly fetch env variables (#64)

* (mod/ci)tweaks to docker.yml

- runs on push and pull-requests
- conditionnally pushes to dockerhub either prod or dev image (only on
  pushes)
- refactoring the way with get app version

* (mod/config)bumping version+using fetch_env! instead of get_env

Release configuration guide:
https://hexdocs.pm/mix/1.14/Mix.Tasks.Release.html#module-vm-args-and-env-sh-env-bat

* (mod/build)docker-compose now relies on prod image version

should be better for tests & stuff
-> now has hardcoded values in environment instead of using .env file
because it might have conflicts

* (mod/build)enabling network_mode: host

since im running WSL, docker wont fetch updates properly when building
docker images, so network_mode: host has to be enabled

* (fix/ci)setting mandatory ENV variables

because we're now using fetch_env!, and error will be triggered when
creating an image, or running the tests, which is why we set default env
variables for the test run

* (add/build)dev docker-compose

to help the mobile & front-end team, we provide a dev docker-compose so
that they can still do their development while enjoying all the perks of
MIX_ENV=dev

to run, just do:

> docker-compose -f docker-compose-dev.yml up --build

* (mod/docs)specifying how to run tests & run dev env

* (fix/ci)interpolating env values to get correct image name

relevant SO thread:
> https://stackoverflow.com/a/72598237

* (fix/ci)APP_VERSION wrapped everywhere

* (mod/ci)pushing dev image with :dev tag along with [version]-dev

* (mod/api)account to domain foreign keys (#65)

* (mod/naboo)adding foreign keys between accounts & nodes

this is the first part of the work we have to do to have nodes &
addresses actually being stored by users. next up:

- encode user id in jwt -> when they create a node, cast their user_id
  with it
(figuring out a way to do this for graphQL as well would be nice, but it
is not mandatory at this point)
- edit changesets to cast the association
- add functions to query all of the necessary data depending on a user
  id, but not touch the current one because it might slow down
  performances overall
- edit & fix tests to reflect these changes

* (mod/ci)specifying DATABASE_SSL & STATIC_URL

AFAIK, these values are not _actually_ used for now, but the work done
to fix release mode has imposed us to use fetch_env! for everything,
which breaks startup of the app is the env value is not set

* (fix/test)specifying DATABASE_POOL_SIZE

* (fix/api)improving account creation error checking

-> changed the way changeset return errors when changeset is invalid, so
that API is more consistent
-> enforcing unique constraints on account creation (not as a business
rule)
-> removing password_confirmation field from account creation, as it
should be done in the front-end
-> checking wether email matches basic email regex

* (mod/build)fixing env values in prod docker-compose

* (mod/api)refactoring views for preloaded data

To not repeat code, we can just call a view from another one, when we
know there is a preloaded query. this is much cleaner, i like this
commit very much.

* (mod/api)setting up foreign-key relationships between nodes & accounts

we can now query an account with its related nodes, this is still
untested, but we need to change some stuff in the node_controller before
testing this part of the feature

* (mod/api)fixed assocs between users & domains


Co-authored-by: winds0r <w.nds0r@protonmail.com>